# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aelu API
  version: 1.0.0
  description: >
    REST API for the Aelu learning system. Supports adaptive drill sessions
    (via WebSocket), spaced-repetition progress tracking, graded reading, media
    exposure, extensive listening, vocabulary encounter logging, xAPI/Caliper
    learning-record export, Common Cartridge export, GDPR data portability and
    erasure, JWT-based mobile auth, offline sync, Stripe payments, and push
    notifications.
servers:
  - url: https://aelu.app
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /api/settings/anonymous-mode:
    get:
      tags: [Settings]
      summary: Get anonymous learning mode setting
      operationId: getAnonymousMode
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  anonymous_mode:
                    type: boolean
    post:
      tags: [Settings]
      summary: Toggle anonymous learning mode
      operationId: setAnonymousMode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  anonymous_mode:
                    type: boolean
        "403":
          description: Feature not available for this user
  /api/settings/marketing-opt-out:
    get:
      tags: [Settings]
      summary: Get marketing email opt-out setting
      operationId: getMarketingOptOut
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  marketing_opt_out:
                    type: boolean
    post:
      tags: [Settings]
      summary: Toggle marketing email opt-out
      operationId: setMarketingOptOut
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                opted_out:
                  type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  marketing_opt_out:
                    type: boolean

  # ── Push Notifications ──────────────────────────────────────────────
  /api/push/register:
    post:
      tags: [Push]
      summary: Register a push notification token
      operationId: registerPush
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, token]
              properties:
                platform:
                  type: string
                  enum: [ios, android]
                token:
                  type: string
                  description: Device push token
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "400":
          description: Missing or invalid platform/token
  /api/push/unregister:
    post:
      tags: [Push]
      summary: Remove push notification token on logout
      operationId: unregisterPush
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                platform:
                  type: string
                  enum: [ios, android]
                  description: If omitted, removes all tokens for the user
      responses:
        "200":
          description: Token removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"

  # ── Personalization ────────────────────────────────────────────────
...
  /api/settings/session-length:
    get:
      tags: [Settings]
      summary: Get preferred session length
      operationId: getSessionLength
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
    post:
      tags: [Settings]
      summary: Set preferred session length
      operationId: setSessionLength
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
  /api/settings/audio:
    get:
      tags: [Settings]
      summary: Get audio playback preferences
      operationId: getAudioSettings
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
    post:
      tags: [Settings]
      summary: Set audio playback preferences
      operationId: setAudioSettings
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
  /api/settings/daily-goal:
    get:
      tags: [Settings]
      summary: Get daily session goal
      operationId: getDailyGoal
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
    post:
      tags: [Settings]
      summary: Set daily session goal
      operationId: setDailyGoal
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
  /api/settings/streak-reminders:
    get:
      tags: [Settings]
      summary: Get streak reminder preferences
      operationId: getStreakReminders
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
    post:
      tags: [Settings]
      summary: Set streak reminder preferences
      operationId: setStreakReminders
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
  /api/settings/display-prefs:
    get:
      tags: [Settings]
      summary: Get display preferences (pinyin visibility, font size, etc.)
      operationId: getDisplayPrefs
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
    post:
      tags: [Settings]
      summary: Set display preferences
      operationId: setDisplayPrefs
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Flask session cookie set on login
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/auth/token
  schemas:
    StatusOk:
      type: object
      properties:
        status:
          type: string
          example: ok
...